# src/openapi.json
{
  "openapi": "2.8.0",
  "info": {
    "title": "XTE Apis",
    "author": "D2:业",
    "version": "v1.0.0",
    "description": "Simple and easy to use API."
  },
  "servers": [
    {
      "url": "https://api.xte.web.id"
    }
  ],
...
  "paths": {
...
    "/ai/copilot-voice": {
      "get": {
        "summary": "Copilot Voice AI",
        "description": "Mendapatkan jawaban AI dalam format teks dan base64 audio.",
        "tags": ["AI"],
        "parameters": [{ "name": "text", "in": "query", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "OK" } }
      }
    },
...
    "/ai/zonerai": {
      "get": {
        "summary": "AI Text to Image (Zonerai)",
        "description": "Generator gambar AI dengan berbagai pilihan aspek rasio (size).",
        "tags": ["AI"],
        "parameters": [
          { "name": "prompt", "in": "query", "required": true, "schema": { "type": "string" } },
          { "name": "size", "in": "query", "required": false, "schema": { "type": "string", "default": "1024x1024", "enum": ["1024x1024", "1216x832", "1152x896", "1344x768", "1563x640", "832x1216", "896x1152", "768x1344", "640x1536"] } }
        ],
        "responses": { "200": { "description": "Image PNG" } }
      }
    },
...
    "/anime/otakudesu-search": {
      "get": {
        "summary": "Search Anime Otakudesu",
        "description": "Mencari judul anime pada platform Otakudesu.",
        "tags": ["Anime"],
        "parameters": [{ "name": "query", "in": "query", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "Success" } }
      }
    },
...
    "/v1/download/gitclone": {
      "get": {
        "summary": "GitHub Repository Downloader",
        "description": "Download repository GitHub dalam bentuk file ZIP secara langsung.",
        "tags": ["Downloader"],
        "parameters": [{ "name": "url", "in": "query", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "Success" } }
      }
    },
    "/v1/download/xv": {
      "get": {
        "summary": "Public Video Downloader (Xham/XV)",
        "description": "Downloader video untuk platform publik (Xhamster/XVideos).",
        "tags": ["Downloader"],
        "parameters": [{ "name": "url", "in": "query", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "Success" } }
      }
    },
...
    "/download/mediafire": {
      "get": {
        "summary": "Mediafire Downloader",
        "description": "Mendapatkan link download dan metadata dari file Mediafire.",
        "tags": ["Downloader"],
        "parameters": [{ "name": "url", "in": "query", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "Success" } }
      }
    },
...
    "/v2/download/ytmp4": {
      "get": {
        "summary": "YouTube MP4 Downloader",
        "description": "Versi cepat. Cukup masukkan link, sistem otomatis memberikan kualitas video terbaik.",
        "tags": ["Downloader"],
        "parameters": [{ "name": "url", "in": "query", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "Success" } }
      }
    },
...
  }
}